requirements.txt
numpy==1.26.4
pandas==2.2.2
TA-Lib==0.4.32
binance-connector==3.8.1
python-dotenv==1.0.1